add sync favorites with db, start to use login page
