+++ src/components/AddToFavorite/AddToFavorite.tsx
@@ -1,7 +1,12 @@
 import styles from './AddToFavorite.module.scss';
 import React from 'react';
-import { addToFavorites, removeFromFavorites } from '../../store/favoriteSlice';
-import { useDispatch } from 'react-redux';
+import {
+  addToFavorites,
+  removeFromFavorites,
+  selectFavoritesProducts,
+} from '../../store/favoriteSlice';
+import { useAppDispatch, useAppSelector } from '../../store/hooks';
+import { SyncUserDataWithServer } from '../../utils/helpers/SyncUserDataWithServer';
 
 type Props = {
   added?: boolean;
@@ -9,14 +14,23 @@ type Props = {
 };
 
 const AddToFavorite: React.FC<Props> = ({ added = false, id }) => {
-  const dispatch = useDispatch();
+  const dispatcher = useAppDispatch();
+  const favoritesInBrowser = useAppSelector(selectFavoritesProducts);
 
   const handleClickAdd = () => {
-    dispatch(addToFavorites(id));
+    dispatcher(addToFavorites(id));
+    const favorites = JSON.stringify({
+      favorite: [...favoritesInBrowser, id],
+    });
+    SyncUserDataWithServer(favorites, 'favorites');
   };
 
   const handleClickRemove = () => {
-    dispatch(removeFromFavorites(id));
+    dispatcher(removeFromFavorites(id));
+    const favorites = JSON.stringify({
+      favorite: favoritesInBrowser.filter(elem => elem !== id),
+    });
+    SyncUserDataWithServer(favorites, 'favorites');
   };
 
   return (
