+++ src/index.tsx
@@ -24,8 +24,8 @@ createRoot(document.getElementById('root') as HTMLDivElement).render(
           <Route path="/" element={<App />}>
             <Route index element={<HomePage />} />
             <Route path="catalog/:type" element={<ProductPage />} />
-            {/* <Route path="catalog/:tablets" element={<ProductPage />} />
-            <Route path="catalog/:accessories" element={<ProductPage />} /> */}
+            <Route path="register" element={<AuthRoot />} />
+            <Route path="login" element={<AuthRoot />} />
 
             <Route path="/cart" element={<CartPage />} />
             <Route path="/favourites" element={<FavouritesPage />} />
