+++ src/types/User.ts
@@ -1,0 +1,7 @@
+'use strict';
+
+export type User = {
+  id: string;
+  email: string;
+  name: string;
+};
